MR-2669 - renamed
